Further deltarpm fixes.
Allow lists of directories to delta against, either via -d command line
option, or delta_dirs config file option. Default to any directory passed
in with -p if neither is set.

Copy in prior deltas from the previous directory if they still match
the package set, so that we don't lose all deltas if we recompose with
the same content.

Pass the delta flag on the second post-multilib createrepo call as well,
to make sure that the metadata ends up in repomd.xml.