Don't suppress error messages
If a failure occured when trying to create the cache folder, mash would
just silently ignore the error and happily continue cranking along. It
might of course fail later on, but the problem is impossible to debug if
errors are silently suppressed.

However, if the error is merely that the directory already exists (as
opposed to a permission issue, for example), then that is fine to
ignore.

Likewise, if a failure occurs when trying to download, merely saying
"can't download" is not helpful. Adding the actual error message makes
it easier to figure out what went wrong.